fix(db): typeorm config not loaded from config service
