+++ src/config/envs/development.ts
@@ -1,8 +1,8 @@
-export const config = () => ({
+export const config = {
   app: {
     environment: 'development',
   },
   db: {
     logging: true,
   },
-});
+};
