--- src/config/envs/index.ts
@@ -1,20 +1,0 @@
-import { getEnv, mergeDeepRight } from '@/utils/helpers';
-
-import { config as base } from './default';
-import { config as custom } from './production';
-
-type TDefaultConfig = typeof base;
-type TCustomConfig = typeof custom;
-export type TConfig = TDefaultConfig & TCustomConfig;
-
-export const configuration = async (): Promise<TConfig> => {
-  const { config: baseConfig }: { config: TDefaultConfig } = await import(
-    `${__dirname}/default`
-  );
-
-  const { config: customConfig }: { config: TCustomConfig } = await import(
-    `${__dirname}/${getEnv()}`
-  );
-
-  return mergeDeepRight(baseConfig, customConfig);
-};
