+++ src/config/index.ts
@@ -1,1 +1,22 @@
-export * from './envs';
+import { getEnv, mergeDeepRight } from '@/utils/helpers';
+
+import { config as base } from './envs/default';
+import { config as custom } from './envs/production';
+
+type TDefaultConfig = typeof base;
+type TCustomConfig = typeof custom;
+export type TConfig = TDefaultConfig & TCustomConfig;
+
+export const configuration = async (): Promise<TConfig> => {
+  const { config: baseConfig }: { config: TDefaultConfig } = await import(
+    `${__dirname}/envs/default`
+  );
+
+  const { config: customConfig }: { config: TCustomConfig } = await import(
+    `${__dirname}/envs/${getEnv()}`
+  );
+
+  const final = mergeDeepRight(baseConfig, customConfig);
+  console.log('final:', final);
+  return final;
+};
